+++ src/test/TestUSDCSubmitOffer.t.sol
@@ -66,7 +66,7 @@ contract TestUSDCSubmitOffer is Setup {
             address(repoToken1Week),
             idHash,
             100,
-            bytes32("test price"),
+            uint256(bytes32("test price")),
             offerAmount
         );
 
@@ -76,7 +76,7 @@ contract TestUSDCSubmitOffer is Setup {
             address(repoToken1Week),
             idHash,
             100,
-            bytes32("test price"),
+            uint256(bytes32("test price")),
             offerAmount
         );
 
@@ -112,7 +112,7 @@ contract TestUSDCSubmitOffer is Setup {
             address(repoToken1Week),
             offerId1,
             100,
-            bytes32("test price"),
+            uint256(bytes32("test price")),
             offerAmount
         );
 
@@ -149,7 +149,7 @@ contract TestUSDCSubmitOffer is Setup {
             address(repoToken1Week),
             offerId1,
             100,
-            bytes32("test price"),
+            uint256(bytes32("test price")),
             60e6
         );
 
@@ -160,7 +160,7 @@ contract TestUSDCSubmitOffer is Setup {
             address(repoToken1Week),
             offerId1,
             100,
-            bytes32("test price"),
+            uint256(bytes32("test price")),
             40e6
         );
     }
